+++ Airship/AirshipCore/Source/AirshipEmbeddedView.swift
@@ -13,7 +13,7 @@ public struct AirshipEmbeddedView<PlaceHolder: View>: View {
     private var viewModel: EmbeddedViewModel
 
     private let placeholder: () -> PlaceHolder
-    private let id: String
+    private let embeddedID: String
     private let embeddedSize: AirshipEmbeddedSize?
 
     /// Creates a new AirshipEmbeddedView.
@@ -22,15 +22,16 @@ public struct AirshipEmbeddedView<PlaceHolder: View>: View {
     ///   - id: The embedded ID.
     ///   - size: The embedded size info. This is needed in a scroll view to determine proper percent based sizing.
     ///   - placeholder: The place holder block.
+    @available(*, deprecated,  message: "use init(embeddedID:embeddedSize:placeHolder:) instead")
     public init(
         id: String,
         embeddedSize: AirshipEmbeddedSize? = nil,
         @ViewBuilder placeholder: @escaping () -> PlaceHolder = { EmptyView()}
     ) {
-        self.id = id
+        self.embeddedID = id
         self.embeddedSize = embeddedSize
         self.placeholder = placeholder
-        self._viewModel = StateObject(wrappedValue: EmbeddedViewModel(id: id))
+        self._viewModel = StateObject(wrappedValue: EmbeddedViewModel(embeddedID: id))
     }
 
     /// Creates a new AirshipEmbeddedView.
@@ -38,14 +39,47 @@ public struct AirshipEmbeddedView<PlaceHolder: View>: View {
     /// - Parameters:
     ///   - id: The embedded ID.
     ///   - size: The embedded size info. This is needed in a scroll view to determine proper percent based sizing.
+    @available(*, deprecated,  message: "use init(embeddedID:embeddedSize:) instead")
     public init(
         id: String,
         embeddedSize: AirshipEmbeddedSize? = nil
     ) where PlaceHolder == EmptyView {
-        self.id = id
+        self.embeddedID = id
         self.embeddedSize = embeddedSize
         self.placeholder = { EmptyView() }
-        self._viewModel = StateObject(wrappedValue: EmbeddedViewModel(id: id))
+        self._viewModel = StateObject(wrappedValue: EmbeddedViewModel(embeddedID: id))
+    }
+
+    /// Creates a new AirshipEmbeddedView.
+    ///
+    /// - Parameters:
+    ///   - embeddedID: The embedded ID.
+    ///   - size: The embedded size info. This is needed in a scroll view to determine proper percent based sizing.
+    ///   - placeholder: The place holder block.
+    public init(
+        embeddedID: String,
+        embeddedSize: AirshipEmbeddedSize? = nil,
+        @ViewBuilder placeholder: @escaping () -> PlaceHolder = { EmptyView()}
+    ) {
+        self.embeddedID = embeddedID
+        self.embeddedSize = embeddedSize
+        self.placeholder = placeholder
+        self._viewModel = StateObject(wrappedValue: EmbeddedViewModel(embeddedID: embeddedID))
+    }
+
+    /// Creates a new AirshipEmbeddedView.
+    ///
+    /// - Parameters:
+    ///   - embeddedID: The embedded ID.
+    ///   - size: The embedded size info. This is needed in a scroll view to determine proper percent based sizing.
+    public init(
+        embeddedID: String,
+        embeddedSize: AirshipEmbeddedSize? = nil
+    ) where PlaceHolder == EmptyView {
+        self.embeddedID = embeddedID
+        self.embeddedSize = embeddedSize
+        self.placeholder = { EmptyView() }
+        self._viewModel = StateObject(wrappedValue: EmbeddedViewModel(embeddedID: embeddedID))
     }
 
     public var body: some View {
@@ -82,10 +116,10 @@ private class EmbeddedViewModel: ObservableObject {
     private var timer: AnyCancellable?
     private var viewManager: AirshipEmbeddedViewManager
     
-    init(id: String, manager: AirshipEmbeddedViewManager = AirshipEmbeddedViewManager.shared) {
+    init(embeddedID: String, manager: AirshipEmbeddedViewManager = AirshipEmbeddedViewManager.shared) {
         self.viewManager = manager
         cancellable = viewManager
-            .publisher(embeddedViewID: id)
+            .publisher(embeddedViewID: embeddedID)
             .receive(on: DispatchQueue.main)
             .sink(receiveValue: onNewViewReceived)
     }